Fix bug on mariadb
Bug on Docker for Mac when we change the database from mysql to mariadb.

Maybe It is not a perfect fix. This fix will not share the same mysql volume, means we could not change database directly. We have to import the database to both database. But it will fix the Bug . Wish for better solution